[+] Implement v.1.0.0 of Mastermind game
This is a massive commit, becausre it's the first stable version of the game. Essentially making the game from scratch without any commits in between. Felt I had to do it this way because, there would be a LOT of refactoring of classes and objects since I'm not very familiar with the gosu library. So if I were to commit, reverting back  would be a bit useless if that implimentation is inferior and it's not a stable version anyway.